requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
python-dotenv>=1.0
pydantic>=2.7
httpx>=0.27
google-api-python-client>=2.130
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
google-genai>=1.0.0
pandas>=2.2
matplotlib>=3.9
streamlit>=1.36
numpy>=1.26
scikit-learn>=1.5
rich>=13.7
langgraph>=0.2
langchain-core>=0.2
chromadb>=0.5.0
